move cards into sidebar component
