+++ src/components/data-global-sidebar/biodiversity-sidebar-card/biodiversity-sidebar-card.js
