+++ src/components/data-global-sidebar/data-global-sidebar-component.jsx
@@ -2,9 +2,9 @@
 import React from 'react';
 import cx from 'classnames';
 // components
-import BiodiversitySidebarCard from 'components/biodiversity-sidebar-card';
-import ProtectedAreasSidebarCard from 'components/protected-areas-sidebar-card';
-import HumanImpactSidebarCard from 'components/human-impact-sidebar-card';
+import BiodiversitySidebarCard from './biodiversity-sidebar-card';
+import ProtectedAreasSidebarCard from './protected-areas-sidebar-card';
+import HumanImpactSidebarCard from './human-impact-sidebar-card';
 import FindPlacesCard from 'components/find-places-card';
 // styles
 import styles from './data-global-sidebar-styles.module.scss';
@@ -32,13 +32,13 @@ const DataGlobalSidebarComponent = ({
         className={styles.biodiversitySidebarCard}
         countedActiveLayers={countedActiveLayers}
       />
-      <HumanImpactSidebarCard
+      <ProtectedAreasSidebarCard
         activeLayers={activeLayers}
         activeCategory={activeCategory}
         handleGlobeUpdating={handleGlobeUpdating}
         countedActiveLayers={countedActiveLayers}
       />
-      <ProtectedAreasSidebarCard
+      <HumanImpactSidebarCard
         activeLayers={activeLayers}
         activeCategory={activeCategory}
         handleGlobeUpdating={handleGlobeUpdating}
