+++ src/components/data-global-sidebar/human-impact-sidebar-card/human-impact-sidebar-card-styles.module.scss
