+++ src/components/data-global-sidebar/protected-areas-sidebar-card/protected-areas-sidebar-card-styles.module.scss
