+++ src/styles/settings.scss
@@ -143,9 +143,9 @@ $high-sidebar-top-margin: 268px;
 $sidebar-header-height: 100px;
 $sidebar-header-height-mobile: 150px;
 $ui-left-top: calc(#{$sidebar-top-margin} + #{$he-project-logo-size});
-$sidebar-width: 350px; // use this for entry boxes and sidebar width
+$sidebar-width: 420px; // use this for entry boxes and sidebar width
 $country-sidebar-width: 490px;
-$local-sidebar-width: 350px;
+$local-sidebar-width: 420px;
 $widgetsTopMargin: 8px;
 $cards-bottom-margin: 8px;
 $border-radius-desktop: 2px;
